vmbus_server: Check if RequestOffers was received before sending offers during restore (#1681)
If the VMBus server state is saved after the guest connects but before
it sends `RequestOffers`, all the channels are in the `ClientReleased`
state and are therefore not saved.

When the saved state is restored, all channels are treated as new.
Because the guest has connected, `revoke_unclaimed_channels` would send
offer messages for these channels (which the guest does not expect at
this point). When the guest subsequently sends `RequestOffers`, these
channels are no longer in the `ClientReleased` state causing an assert
to fail and OpenVMM/OpenHCL panics.

This change fixes `revoke_unclaimed_channels` to only send offers if
`RequestOffers` was received before the save, and adds tests to cover
both scenarios.